extend fix to all files

+++ devise.rb
@@ -53,11 +53,13 @@ run 'rm -rf vendor'
 run 'curl -L https://github.com/lewagon/stylesheets/archive/master.zip > stylesheets.zip'
 run 'unzip stylesheets.zip -d app/assets && rm stylesheets.zip && mv app/assets/rails-stylesheets-master app/assets/stylesheets'
 
-run 'rm app/assets/javascripts/application.js'
-file 'app/assets/javascripts/application.js', <<-JS
+if Rails.version < 6
+  run 'rm app/assets/javascripts/application.js'
+  file 'app/assets/javascripts/application.js', <<-JS
 //= require rails-ujs
 //= require_tree .
-JS
+  JS
+end
 
 # Dev environment
 ########################################
@@ -82,7 +84,7 @@ file 'app/views/layouts/application.html.erb', <<-HTML
     <%= render 'shared/navbar' %>
     <%= render 'shared/flashes' %>
     <%= yield %>
-    <%= javascript_include_tag 'application' %>
+#{"    <%= javascript_include_tag 'application' %>\n" if Rails.version < 6}
     <%= javascript_pack_tag 'application' %>
   </body>
 </html>
